requirements.txt
streamlit
streamlit.lottie
google-cloud-firestore